requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch==1.13.1
torchvision==0.14.1
flake8
Click
loguru
packaging
sentencepiece
peft==0.4.0
einops==0.6.1
gradio==3.39.0
markdown2==2.4.10
numpy==1.23.5
opencv_python==4.8.0.74
Pillow==9.4.0
pycocotools==2.0.6
shortuuid==1.0.11
tqdm==4.64.1
transformers==4.33.2
uvicorn==0.23.2
scipy
bitsandbytes==0.41.1
timm==0.6.13
tabulate==0.9.0
omegaconf==2.3.0
addict
pytest
bleach
requests
scikit-image
deepspeed==0.9.5
langchain==0.0.311
# add sd decoder
diffusers==0.21.2